RUN: Patch a command line to make it runnable in a remote context
